Refactor memory-mapping interfaces in sanitizer_common.h.
It makes more sense, semantically, for most callers of
MmapNoAccess and MmapFixed, to instead create a "reserved
address range" object, instead of trying to map memory
while marking it inaccessible.

In particular, this interface is necessary for Fuchsia,
which requires a handle to interact with the address space,
rather than arbitrary addresses.

So, we create a ReservedAddressRange object & a logical
interface for it here.

Change-Id: I8cc2960ece69c84b57de07c826ef30bbed28d1f9